Fixing additional signedness errors after adding -Wsign-conversion flag

+++ libsmtutil/Z3Interface.cpp
@@ -184,7 +184,7 @@ z3::expr Z3Interface::toZ3Expr(Expression const& _expr)
 			return arguments[0] & arguments[1];
 		else if (n == "int2bv")
 		{
-			size_t size = std::stoi(_expr.arguments[1].name);
+			size_t size = std::stoul(_expr.arguments[1].name);
 			return z3::int2bv(size, arguments[0]);
 		}
 		else if (n == "bv2int")
